+++ tests/test_idempotency_keys.py
@@ -20,10 +20,10 @@ class IdempotencyKeyTest(TestCase):
         key4 = get_idempotency_key("charge", "create:1", False)
         self.assertTrue(key1 != key4)
 
-        self.assertEquals(IdempotencyKey.objects.count(), 3)
+        self.assertEqual(IdempotencyKey.objects.count(), 3)
         key1_obj = IdempotencyKey.objects.get(action="customer:create:1", livemode=False)
         self.assertFalse(key1_obj.is_expired)
-        self.assertEquals(str(key1_obj), str(key1_obj.uuid))
+        self.assertEqual(str(key1_obj), str(key1_obj.uuid))
 
     def test_clear_expired_idempotency_keys(self):
         expired_key = get_idempotency_key("customer", "create:1", False)
@@ -33,9 +33,9 @@ class IdempotencyKeyTest(TestCase):
 
         valid_key = get_idempotency_key("customer", "create:2", False)
 
-        self.assertEquals(IdempotencyKey.objects.count(), 2)
+        self.assertEqual(IdempotencyKey.objects.count(), 2)
 
         clear_expired_idempotency_keys()
 
-        self.assertEquals(IdempotencyKey.objects.count(), 1)
-        self.assertEquals(str(IdempotencyKey.objects.get().uuid), valid_key)
+        self.assertEqual(IdempotencyKey.objects.count(), 1)
+        self.assertEqual(str(IdempotencyKey.objects.get().uuid), valid_key)