+++ HISTORY.rst
@@ -39,6 +39,7 @@ History
 * Implemented ``Subscription.reactivate()`` (Thanks @jleclanche) #470
 * Added ``Plan.human_readable_price`` (Thanks @jleclanche) #498
 * (Re)attach the Subscriber when we find it's id attached to a customer on Customer sync (Thanks @jleclanche) #500
+* Made API version configurable (with dj-stripe recommended default) (Thanks @lskillen) #504
 
 
 0.8.0 (2015-12-30)